fixes from review - restructured java code and removed annotations
